packages/metaed-plugin-edfi-ods-postgresql/test/generator/SchemaGenerator.test.ts
import { newMetaEdEnvironment, newNamespace } from 'metaed-core';
import { GeneratorResult, MetaEdEnvironment, Namespace } from 'metaed-core';
import { generate } from '../../src/generator/SchemaGenerator';

describe('when generating schemas for core namespace', (): void => {
  let result: GeneratorResult;

  beforeAll(async () => {
    const metaEd: MetaEdEnvironment = newMetaEdEnvironment();
    const namespace: Namespace = {
      ...newNamespace(),
      namespaceName: 'EdFi',
      isExtension: false,
    };
    metaEd.namespace.set('EdFi', namespace);

    result = await generate(metaEd);
  });

  it('should generate correct schema', (): void => {
    expect(result.generatorName).toEqual('edfiOdsPostgresql.SchemaGenerator');
    expect(result.generatedOutput[0].fileName).toBe('0010-Schemas.sql');
    expect(result.generatedOutput[0].namespace).toBe('EdFi');
    expect(result.generatedOutput[0].folderName).toBe('/Database/PostgreSQL/ODS/Structure/');
    expect(result.generatedOutput[0].name).toBe('ODS PostgreSQL Schema');
    expect(result.generatedOutput[0].resultStream).toBeNull();
    expect(result.generatedOutput[0].resultString).toMatchInlineSnapshot(`
      "CREATE SCHEMA auth AUTHORIZATION postgres;
      CREATE SCHEMA edfi AUTHORIZATION postgres;
      CREATE SCHEMA util AUTHORIZATION postgres;
      "
    `);
  });
});

describe('when generating schemas for extension namespace', (): void => {
  let result: GeneratorResult;

  beforeAll(async () => {
    const metaEd: MetaEdEnvironment = newMetaEdEnvironment();
    const namespace: Namespace = {
      ...newNamespace(),
      namespaceName: 'Extension',
      projectExtension: 'EXTENSION',
      isExtension: true,
    };
    metaEd.namespace.set('Extension', namespace);

    result = await generate(metaEd);
  });

  it('should generate correct schema', (): void => {
    expect(result.generatorName).toEqual('edfiOdsPostgresql.SchemaGenerator');
    expect(result.generatedOutput[0].fileName).toBe('0010-EXTENSION-Extension-Schemas.sql');
    expect(result.generatedOutput[0].namespace).toBe('Extension');
    expect(result.generatedOutput[0].folderName).toBe('/Database/PostgreSQL/ODS/Structure/');
    expect(result.generatedOutput[0].name).toBe('ODS PostgreSQL Schema');
    expect(result.generatedOutput[0].resultStream).toBeNull();
    expect(result.generatedOutput[0].resultString).toMatchInlineSnapshot(`
      "CREATE SCHEMA extension AUTHORIZATION postgres;
      "
    `);
  });
});
